requirements.txt
# Core Python packages and ML frameworks
pytest
transformers
numpy
pandas
torch==2.5.1
tensorflow==2.17
tf-keras==2.17
scipy==1.14.1
PyYAML
regex

# Presidio PII Anonymization components
presidio-anonymizer>=2.2.2,<2.3
presidio-analyzer>=2.2,<2.3

# Logging and monitoring
prometheus_client==0.22.1
python-json-logger

# API, server, and cloud integration
fastapi
uvicorn
boto3
slack_sdk
google-cloud-storage
python-dotenv

# Language and NLP utilities
flair~=0.14
fasttext
huggingface-hub
pydantic

# spaCy models (language-specific)
hu_core_news_lg@https://huggingface.co/huspacy/hu_core_news_lg/resolve/main/hu_core_news_lg-any-py3-none-any.whl
de_core_news_lg@https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl
el_core_news_lg@https://github.com/explosion/spacy-models/releases/download/el_core_news_lg-3.8.0/el_core_news_lg-3.8.0-py3-none-any.whl

# File I/O
openpyxl

# Local wheels (language identification)
./required_wheels/languageidentifier-1.3-py3-none-any.whl
